profile pic done

+++ backend/src/server.ts
@@ -5,6 +5,7 @@ import { connectToDB } from "./db/conectdb";
 import bodyParser from "body-parser";
 import morgan from "morgan";
 import cors from "cors";
+import fileupload from "express-fileupload";
 
 import userRotes from "./routes/user.registration.route";
 import gameRoutes from "./routes/game.route";
@@ -20,6 +21,7 @@ import onlineUsersRoute from "./routes/OnlineUsers.route";
 const app = express();
 const PORT = config.get("PORT");
 
+app.use(fileupload({ useTempFiles: true, tempFileDir: "/temp/" }));
 app.use(cors());
 app.use(express.urlencoded({ extended: false }));
 app.use(morgan("dev"));
@@ -36,9 +38,9 @@ const server = app.listen(PORT, () => {
   console.log("Server is Running");
 });
 
-const socketIo = init(server);
+const socketIo: any = init(server);
 
-socketIo.use(async (socket, next) => {
+socketIo.use(async (socket: any, next: any) => {
   const walletId = socket.handshake.auth.walletId;
   const checkUser = await userModal.findOne({ wallet: walletId });
   console.log({ checkUser, walletId });
@@ -54,6 +56,7 @@ socketIo.use(async (socket, next) => {
       name: "",
       socketId: socket.id,
       wallet: walletId,
+      profilePic: "",
     });
     await newUser.save();
     console.log({ newUser });
@@ -67,12 +70,7 @@ socketIo.on("connection", async (socket: any) => {
   socket.emit("message", `connected with id ${socket.id}`);
 
   socket.on("message", (msg: any) => {
-    console.log(msg);
-  });
-
-  socket.on("connecteToOppoent", (opponentId: string) => {
-    socket.join(opponentId);
-    socket.emit("msgFromOpponent", "Hello World");
+    console.log({ msg });
   });
 
   socket.on("disconnect", () => {
@@ -80,22 +78,32 @@ socketIo.on("connection", async (socket: any) => {
     removeUser(socket.id);
   });
 
-  socket.on("availableForMactch", async (walletId: any, amount: any) => {
-    console.log("Player Available for Match", walletId);
-
-    const checkUser = await userModal.findOne({ wallet: walletId });
-    if (checkUser) {
-      const newEntry = new availableUserModel({
-        // adding current user to available list
-        amount: amount,
-        userId: checkUser._id,
-        socketId: checkUser.socketId,
-      });
-      await newEntry.save();
-
-      startMatching(socket, newEntry);
-    }
+  socket.on("msgToCustomRoom", (msg: string, roomId: string) => {
+    // msg to custome room
+    socket.in(roomId).emit("message", msg);
   });
+
+  socket.on(
+    "availableForMatch",
+    async (walletId: any, amount: any, level: any) => {
+      console.log("Player Available for Match", walletId);
+
+      const checkUser = await userModal.findOne({ wallet: walletId });
+      if (checkUser) {
+        const newEntry = new availableUserModel({
+          // adding current user to available list
+          amount: amount,
+          level: level,
+          userId: checkUser._id,
+          socketId: checkUser.socketId,
+          profilePic: checkUser.profilePic,
+        });
+        await newEntry.save();
+
+        startMatching(socket, newEntry);
+      }
+    }
+  );
 });
 
 async function startMatching(socket: any, data: any) {
@@ -103,25 +111,39 @@ async function startMatching(socket: any, data: any) {
   try {
     console.log("starting Search");
 
-    let count = 0;
+    // let count = 0;
     let isOpponent = false;
 
     while (!isOpponent) {
+      const checkAvailablility = await availableUserModel.find({
+        userId: data.userId,
+        amount: data.amount,
+        level: data.level,
+      });
+
+      if (!checkAvailablility) {
+        return (isOpponent = true);
+      }
+
       const opponent = await availableUserModel.findOne({
         userId: { $ne: data.userId },
         amount: data.amount,
+        level: data.level,
       });
       if (opponent) {
-        socket.in(opponent.socketId).emit("gotOpponent", data);
-        socket.in(data.socketId).emit("gotOpponent", opponent);
-        console.log({ opponent });
-        isOpponent = true;
         await availableUserModel.findOneAndRemove({
           userId: opponent.userId,
         });
         await availableUserModel.findOneAndRemove({ userId: data.userId });
+        socket.to(opponent.socketId).emit("gotOpponent", data, data.socketId);
+        socket
+          .to(data.socketId)
+          .emit("gotOpponent", opponent, opponent.socketId);
+        console.log({ opponent });
+        // addnewGameEntry(data.userId, opponent.userId, data.amount);
+
+        isOpponent = true;
       } else {
-        // await availableUserModel.findOneAndRemove({ userId: data.userId });
         socket.emit("noOpponent", {
           opponent: false,
           message: "no oppenent Found",
